ecryptfs: fix compile warning for 32-bit build
fix compile warning for 32-bit build to avoid build failure.

Reviewed-on: http://gerrit.mot.com/849468
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ fs/ecryptfs/main.c
@@ -467,7 +467,8 @@ static int ecryptfs_parse_options(struct ecryptfs_sb_info *sbi, char *options,
 				mount_crypt_stat->global_default_cipher_name,
 				mount_crypt_stat->global_default_cipher_mode,
 				final, sizeof(final)),
-			mount_crypt_stat->global_default_cipher_key_size);
+				(long unsigned int)
+				mount_crypt_stat->global_default_cipher_key_size);
 		rc = -EINVAL;
 		goto out;
 	}
